Resturcturing ibdb routes/blueprints, tests; renaming for clarity; duplicate code cleanup
