+++ ibdb/cli.py
@@ -2,7 +2,7 @@ import os
 
 from ibdb.api import app
 from ibdb.populate_db import main as populate_db_main
-from ibdb.sync import export as sync_export
+from ibdb.export import export as sync_export
 
 
 @app.cli.command('api')
