+++ ibdb/routes/reference/templates/reference/list_edit.html
@@ -31,7 +31,7 @@
     </style>
 </head>
 <body>
-    {% include '_header.html' %}
+    {% include 'header.html' %}
 
     <h1>Add Reference</h1>
     <form method="POST" action="/references">
